final/requirements.txt
fastapi
uvicorn
boto3
jinja2
starlette
python-multipart
python-dotenv
